+++ server.js
@@ -58,6 +58,7 @@ app.use((req, res, next) => {
 
 // Routes
 
+app.use(require('./routes/admin.routes'));
 app.use(require('./routes/index.routes'));
 app.use(require('./routes/reports.routes'));
 app.use(require('./routes/users.routes'));