+++ CHANGELOG.md
@@ -4,7 +4,26 @@ All notable changes will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.0.1] - 2021-08-29 (Yes, two days after the first version)
+---
+
+## [1.1.0] - 2021-08-30
+
+A lot of new things
+
+### Added
+
+- A more refreshed UI.
+- An admin dashboard.
+
+### Changed
+
+- Nav now float.
+- Type error on `config/passport.js`. 
+- UI changes.
+
+---
+
+## [1.0.1] - 2021-08-29
 
 Only minor things.
 
@@ -20,6 +39,8 @@ Only minor things.
 ### Removed
 
 - Deleted some unnecesary files.
+
+---
 
 ## [1.0.0] - 2021-08-27
 
@@ -37,6 +58,8 @@ The initial release of Green Complaint.
 ### Removed
 
 - The previous code, like ALL that code of Dev versions.
+
+---
 
 ## [0.0.0] - 2021-8-25